modified arragment of MC analysis code
