--- ExoticHiggsDecays_analysis_code/src/classes.h
@@ -1,3 +1,0 @@
-namespace ExoticHiggsDecays_analysis_code {
-  struct dictionary {};
-}  // namespace ExoticHiggsDecays_analysis_code
