--- ExoticHiggsDecays_analysis_code/src/classes_def.xml
@@ -1,3 +1,0 @@
-<lcgdict>
-  <class name="ExoticHiggsDecays_analysis_code::dictionary"/>
-</lcgdict>
