+++ analyzers/dataframe/src/myUtils.cc
@@ -13,6 +13,38 @@ namespace FCCAnalyses{
 
 namespace myUtils{
 
+// following code is added for the Exotic Higgs to LLPs analysis 
+
+ROOT::VecOps::RVec<float> get_both_scalars(ROOT::VecOps::RVec<float> scalar1_value, ROOT::VecOps::RVec<float> scalar2_value) {
+  ROOT::VecOps::RVec<float> sum;
+  for (auto i: scalar1_value) {
+    sum.push_back(i);
+  }
+  for (auto j: scalar2_value) {
+    sum.push_back(j);
+  }
+  return sum;
+}
+
+// following code is copied from https://github.com/HEP-FCC/FCCAnalyses/blob/09c52e107e308cf57d59ca840d4bba60e25d94c2/examples/FCCee/flavour/Bc2TauNu/analysis_B2TauNu_truth.py 
+// used to find the b quarks from respective hs. Can't directly apply sel_byIndex since this function returns a 'edm4hep::MCParticleData' object and we need 'ROOT::VecOps::RVec<edm4hep::MCParticleData>'
+
+selMC_leg::selMC_leg( int arg_idx ) : m_idx(arg_idx) { };
+ROOT::VecOps::RVec<edm4hep::MCParticleData> selMC_leg::operator() ( ROOT::VecOps::RVec<int> list_of_indices,  ROOT::VecOps::RVec<edm4hep::MCParticleData> in) {
+  ROOT::VecOps::RVec<edm4hep::MCParticleData> res;
+  if ( list_of_indices.size() == 0) return res;
+  if ( m_idx < list_of_indices.size() ) {
+	res.push_back( FCCAnalyses::MCParticle::sel_byIndex( list_of_indices[m_idx], in ) );
+	return res;
+  }
+  else {
+	std::cout << "   !!!  in selMC_leg:  idx = " << m_idx << " but size of list_of_indices = " << list_of_indices.size() << std::endl;
+  }
+  return res;
+}
+
+// end of added code for Exotic Higgs to LLPs analysis
+
 float get_d0(TVector3 x, TVector3 p){
   float D = x[1]*p[0]/sqrt(p[0]*p[0]+p[1]*p[1])-x[0]*p[1]/sqrt(p[0]*p[0]+p[1]*p[1]);
   return D;
